+++ src/main/java/sopt/org/HMH/challenge/domain/Challenge.java
@@ -1,4 +1,4 @@
-package sopt.org.HMH.common.domain;
+package sopt.org.HMH.challenge.domain;
 
 import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
@@ -7,6 +7,9 @@ import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import sopt.org.HMH.common.domain.BaseTimeEntity;
+import sopt.org.HMH.dayChallenge.domain.DayChallenge;
+import sopt.org.HMH.user.User;
 
 import java.time.LocalDateTime;
 import java.util.List;
