--- src/main/java/sopt/org/HMH/common/domain/User.java
@@ -1,4 +1,0 @@
-package sopt.org.HMH.common.domain;
-
-public class User extends BaseTimeEntity {
-}
