+++ src/main/java/sopt/org/HMH/user/User.java
@@ -1,0 +1,6 @@
+package sopt.org.HMH.user;
+
+import sopt.org.HMH.common.domain.BaseTimeEntity;
+
+public class User extends BaseTimeEntity {
+}
